pyproject.toml
[project]
name = "ocr-service"
version = "1.0.0"
description = "Multi-engine OCR microservice — Tesseract, Ollama, Surya, EasyOCR, PaddleOCR, DocTR"
requires-python = ">=3.11,<3.12"
dependencies = [
    # ── API framework ──────────────────────────────────────────────────────────
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.12",
    # ── Image & PDF processing ─────────────────────────────────────────────────
    "pillow>=10.4.0",
    "pypdfium2>=4.30.0",
    # ── HTTP (Ollama client) ───────────────────────────────────────────────────
    "httpx>=0.27.0",
    # ── Numeric — kept <2.0 for EasyOCR / PaddleOCR compatibility ─────────────
    "numpy>=1.26.0,<2.0.0",
    # ── OCR engines ───────────────────────────────────────────────────────────
    "pytesseract>=0.3.13",
    "easyocr>=1.7.2",
    "paddlepaddle>=3.0.0",
    "paddleocr>=3.0.0",
    "surya-ocr>=0.17.0",
    "python-doctr[torch]>=0.10.0",
    # ── Deep learning (CPU-only on Linux via [tool.uv.sources]) ───────────────
    "torch>=2.3.1",
    "torchvision>=0.18.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ── uv configuration ──────────────────────────────────────────────────────────
# On Linux (Docker) use the compact CPU-only PyTorch wheels.
# On macOS uv falls back to the standard PyPI torch (which uses MPS/CPU).
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
]
torchvision = [
    { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
]
